Update training.py
Adds stopgaps during the running of the script to exit at certain
points of the script. It also allows you to start training from a
previously preprocessed source.

The script also verifies that the preprocessed files exist before
executing the training, as well as a while section in regards to
choosing a device to do the training with.