adding sample.py
sample.py takes a model and generates a sample set. .gitignore now
ignores the folder generated, and main.py now has the option to use
sample.py. Currently, the script does not write per line to the sample
text file so you would have to keyboard interrupt it for it to be saved.
Will fix that in a later commit. Don't request pull request until this
is fixed.